:art:high_scores: simplify latest
I reread the problem description,
found out that it explicitly mentions list and tuple.
I just read the first paragraph before.
I chose to convert to a deque before
because this works when score is not something sliceable.
However, given that scores is a list or tuple,
converting to a deque unnecessary makes code more complex
and less readable.