:art:python:matrix: np.matrix -> np.array
Thanks for hop to remind me that np.matrix is deprecated.

I ported it to np.array.
Unlike numpy.matrix(string), loadtxt is not smart on number type
(always defaults to float, even when all numbers in the text are integers).
On the other side, loadtxt is smart on dimensions
(which makes sense, since it is not assumed 2d).